requirements.txt
msgpack~=1.0.4
requests~=2.27.1
Flask~=2.1.3
PyYAML~=6.0
UnityPy~=1.9.4
asyncio~=3.4.3
Pillow~=9.1.0
waitress~=2.1.2
python-dateutil~=2.8.2
pydantic~=1.9.0
colorama~=0.4.4
websocket-client~=1.3.2
numpy~=1.22.4
pytz~=2022.1
aiofiles~=0.8.0
wordcloud~=1.8.2.2
jieba~=0.42.1
matplotlib~=3.5.3
configobj~=5.0.6
APScheduler~=3.9.1
html2text~=2020.1.16
xmltodict~=0.13.0
mutagen~=1.45.1
pydub~=0.25.1
zhconv~=1.4.3
aiocqhttp~=1.4.3
aiohttp~=3.7.4.post0
svgwrite~=1.4.3
pymongo~=3.12.3
CairoSVG~=2.5.2
tqdm~=4.64.0
urllib3~=1.26.9
opencv-python~=4.5.5
esprima~=4.0.1
beautifulsoup4~=4.11.1
oss2~=2.16.0
PyMySQL~=1.0.2
ujson~=5.4.0
Werkzeug==2.2.2